+++ src/main/java/com/rockoon/domain/ticket/entity/Ticket.java
@@ -1,19 +1,23 @@
 package com.rockoon.domain.ticket.entity;
 
+import com.rockoon.domain.auditing.entity.BaseTimeEntity;
 import jakarta.persistence.*;
-import lombok.Getter;
-import lombok.Setter;
+import lombok.*;
+import lombok.experimental.SuperBuilder;
 
 import java.util.Date;
 
 @Entity
 @Getter
-@Setter
-public class Ticket {
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@AllArgsConstructor
+@SuperBuilder
+@Table(name = "ticket")
+public class Ticket extends BaseTimeEntity{
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long ticketId;
+    private Long id;
 
     private String data;
     private Date dueDate;
